feat(filter): rename filter group to filter
