+++ packages/paste-website/src/component-examples/filter/components/DefaultFilterGroup.tsx
@@ -22,7 +22,7 @@ import type {
 } from "../types";
 import { EmptyState } from "./EmptyState";
 import { ExtendedDataGrid } from "./ExtendedDataGrid";
-import { FilterPill } from "./FilterPill";
+// import { FilterPill } from "./FilterPill";
 import { SampleDataGrid } from "./SampleDataGrid";
 import { AddFilters } from "./filters/AddFilters";
 import { CustomFilter } from "./filters/CustomFilter";
@@ -34,6 +34,7 @@ import { RoomSidFilter } from "./filters/RoomSidFilter";
 import { RoomTypeFilter } from "./filters/RoomTypeFilter";
 import { SearchFilter } from "./filters/SearchFilter";
 import { UniqueNameFilter } from "./filters/UniqueNameFilter";
+import { FilterPill } from "./FilterPill";
 
 function countMoreFilters(selectedMoreFilters: Record<string, string | string[]>): number {
   let count = 0;
@@ -232,6 +233,7 @@ export const DefaultFilterGroup: React.FC<React.PropsWithChildren<FilterGroupPro
           </Button>
         ) : null}
       </FormPillGroup>
+
 
       <Box display="flex" justifyContent="space-between" alignItems="center" columnGap="space30" marginTop="space50">
         <Box display="flex" columnGap="space30">
