+++ packages/paste-website/src/component-examples/filter/components/FilterPill.tsx
