+++ packages/paste-website/src/component-examples/filter/components/NoResultsIllustration.tsx
