+++ packages/paste-website/src/component-examples/filter/components/filters/MoreFilters.tsx
@@ -103,7 +103,7 @@ const DisclosureFilter = ({
             <Box as="span">{filter.label}</Box>
 
             {selectedCount ? (
-              <Badge as="span" variant="neutral_counter">
+              <Badge as="span" variant="neutral_counter" size="small">
                 Selected {filter.type === "status" ? 1 : selectedCount}
               </Badge>
             ) : null}
