+++ packages/paste-website/src/component-examples/filter/components/filters/ParticipantsFilter.tsx
